Use `std::backtrace::Backtrace` as the default backtrace provider
This also bumps our default Rust version to 1.65.

Closes #357
