--- compatibility-tests/backtrace-shim/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "backtrace-shim"
-version = "0.1.0"
-authors = ["Jake Goulding <jake.goulding@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-snafu = { path = "../..", features = ["backtraces"] }
