--- compatibility-tests/backtraces-impl-std/rust-toolchain
@@ -1,1 +1,0 @@
-beta
