+++ compatibility-tests/report-provider-api/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-snafu = { path = "../..", features = ["backtraces-impl-std", "unstable-provider-api"] }
+snafu = { path = "../..", features = ["unstable-provider-api"] }
