+++ compatibility-tests/v1_56/Cargo.toml
@@ -5,4 +5,4 @@ authors = ["Jake Goulding <jake.goulding@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-snafu = { path = "../..", default-features = false, features = ["backtraces"] }
+snafu = { path = "../..", default-features = false, features = ["std"] }
