+++ compatibility-tests/v1_61/rust-toolchain
@@ -1,0 +1,1 @@
+1.61
