+++ src/guide/feature_flags.md
@@ -7,9 +7,7 @@ cases:
 - [`std`](#std)
 - [`unstable-core-error`](#unstable-core-error)
 - [`guide`](#guide)
-- [`backtraces`](#backtraces)
 - [`backtraces-impl-backtrace-crate`](#backtraces-impl-backtrace-crate)
-- [`backtraces-impl-std`](#backtraces-impl-std)
 - [`unstable-provider-api`](#unstable-provider-api)
 - [`futures`](#futures)
 - [`unstable-try-trait`](#unstable-try-trait)
@@ -63,38 +61,11 @@ built.
 
 Most usages of SNAFU will want this feature disabled.
 
-## `backtraces`
-
-<dl class="snafu-ff-meta">
-<dt>Default</dt>
-<dd>disabled</dd>
-<dt>Implies</dt>
-<dd>
-
-[`std`](#std)
-
-</dd>
-</dl>
-
-When enabled, the [`Backtrace`] type in your enum variant will capture
-a backtrace when the error is generated. If you never use backtraces,
-you can omit this feature to speed up compilation a small amount.
-
-It is recommended that only applications make use of this feature.
-
-[`Backtrace`]: crate::Backtrace
-
 ## `backtraces-impl-backtrace-crate`
 
 <dl class="snafu-ff-meta">
 <dt>Default</dt>
 <dd>disabled</dd>
-<dt>Implies</dt>
-<dd>
-
-[`backtraces`](#backtraces)
-
-</dd>
 </dl>
 
 When enabled, the SNAFU [`Backtrace`] type becomes an alias to the
@@ -102,20 +73,11 @@ When enabled, the SNAFU [`Backtrace`] type becomes an alias to the
 crates that require this type.
 
 It is recommended that only applications make use of this
-feature. When the standard library stabilizes its own backtrace type,
-this feature will no longer be supported and will be removed.
+feature. When SNAFU's minimum supported Rust version has a stable
+`Backtrace` type, this feature will no longer be supported and will be
+removed.
 
-## `backtraces-impl-std`
-
-<dl class="snafu-ff-meta">
-<dt>Default</dt>
-<dd>disabled</dd>
-</dl>
-
-When enabled, the SNAFU [`Backtrace`] type becomes an alias to the
-[`std::backtrace::Backtrace`] type.
-
-It is recommended that only applications make use of this feature.
+[`Backtrace`]: crate::Backtrace
 
 ## `unstable-provider-api`
