+++ tests/premade_error.rs
@@ -103,5 +103,5 @@ fn has_a_backtrace() {
 
     let e = exercise(false).unwrap_err();
     let bt = ErrorCompat::backtrace(&e).expect("Must have a backtrace");
-    assert_eq!("disabled backtrace", bt.to_string());
+    assert!(bt.to_string().contains("has_a_backtrace"));
 }
